POLIO-799: code review
