+++ plugins/polio/js/src/pages/Budget/BudgetDetails/styles.ts
@@ -1,0 +1,67 @@
+import { makeStyles } from '@material-ui/core';
+
+export const useStyles = makeStyles(theme => ({
+    root: {
+        width: '100%',
+        '&.MuiStepper-root': {
+            padding: 0,
+            marginTop: theme.spacing(2),
+        },
+        '& .MuiBox-root': {
+            alignItems: 'flex-start',
+        },
+    },
+    step: {
+        display: 'flex',
+        flexDirection: 'column',
+        alignItems: 'center',
+    },
+    checkboxWrapper: {
+        display: 'flex',
+        alignItems: 'center',
+    },
+    taskIcon: {
+        transform: 'scale(0.85)',
+        position: 'relative',
+        top: -3,
+    },
+    taskDone: {
+        color: theme.palette.success.main,
+    },
+    taskPending: {
+        // @ts-ignore
+        color: theme.palette.mediumGray.main,
+    },
+    itemLabel: {
+        fontSize: '0.82rem',
+    },
+    divider: {
+        height: '30px',
+        // @ts-ignore
+        backgroundColor: theme.palette.gray.main,
+    },
+    stepCompleted: {
+        '& + div .MuiStepConnector-lineHorizontal': {
+            borderColor: theme.palette.success.main,
+        },
+    },
+    stepActive: {
+        '& + div .MuiStepConnector-lineHorizontal': {
+            // @ts-ignore
+            borderColor: theme.palette.yellow.main,
+        },
+    },
+    stepInactive: {
+        '& + div .MuiStepConnector-lineHorizontal': {
+            // @ts-ignore
+            borderColor: theme.palette.gray.main,
+        },
+    },
+    strikethrough: {
+        textDecoration: 'line-through',
+    },
+    overrideIcon: {
+        // @ts-ignore
+        color: theme.palette.gray.main,
+    },
+}));
